requirements.txt
flask
Flask-Babel
fabric3
poxls
